Clean up wiring for the ISS handler.
