+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/state/IssHandlerTests.java
@@ -28,6 +28,7 @@ import com.swirlds.config.api.Configuration;
 import com.swirlds.config.extensions.test.fixtures.TestConfigBuilder;
 import com.swirlds.platform.components.common.output.FatalErrorConsumer;
 import com.swirlds.platform.config.StateConfig_;
+import com.swirlds.platform.state.iss.DefaultIssHandler;
 import com.swirlds.platform.state.iss.IssHandler;
 import com.swirlds.platform.state.iss.IssScratchpad;
 import com.swirlds.platform.system.state.notifications.IssNotification;
@@ -59,7 +60,7 @@ class IssHandlerTests {
 
         final Scratchpad<IssScratchpad> simpleScratchpad = new SimpleScratchpad<>();
         final IssHandler handler =
-                new IssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
+                new DefaultIssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
 
         handler.issObserved(new IssNotification(1234L, IssType.OTHER_ISS));
 
@@ -95,7 +96,7 @@ class IssHandlerTests {
 
         final Scratchpad<IssScratchpad> simpleScratchpad = new SimpleScratchpad<>();
         final IssHandler handler =
-                new IssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
+                new DefaultIssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
 
         handler.issObserved(new IssNotification(1234L, IssType.OTHER_ISS));
 
@@ -125,7 +126,7 @@ class IssHandlerTests {
 
         final Scratchpad<IssScratchpad> simpleScratchpad = new SimpleScratchpad<>();
         final IssHandler handler =
-                new IssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
+                new DefaultIssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
 
         handler.issObserved(new IssNotification(1234L, IssType.SELF_ISS));
 
@@ -157,7 +158,7 @@ class IssHandlerTests {
 
         final Scratchpad<IssScratchpad> simpleScratchpad = new SimpleScratchpad<>();
         final IssHandler handler =
-                new IssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
+                new DefaultIssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
 
         handler.issObserved(new IssNotification(1234L, IssType.SELF_ISS));
 
@@ -189,7 +190,7 @@ class IssHandlerTests {
 
         final Scratchpad<IssScratchpad> simpleScratchpad = new SimpleScratchpad<>();
         final IssHandler handler =
-                new IssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
+                new DefaultIssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
 
         handler.issObserved(new IssNotification(1234L, IssType.SELF_ISS));
 
@@ -227,7 +228,7 @@ class IssHandlerTests {
 
         final Scratchpad<IssScratchpad> simpleScratchpad = new SimpleScratchpad<>();
         final IssHandler handler =
-                new IssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
+                new DefaultIssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
 
         handler.issObserved(new IssNotification(1234L, IssType.CATASTROPHIC_ISS));
 
@@ -259,7 +260,7 @@ class IssHandlerTests {
 
         final Scratchpad<IssScratchpad> simpleScratchpad = new SimpleScratchpad<>();
         final IssHandler handler =
-                new IssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
+                new DefaultIssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
 
         handler.issObserved(new IssNotification(1234L, IssType.CATASTROPHIC_ISS));
 
@@ -297,7 +298,7 @@ class IssHandlerTests {
 
         final Scratchpad<IssScratchpad> simpleScratchpad = new SimpleScratchpad<>();
         final IssHandler handler =
-                new IssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
+                new DefaultIssHandler(platformContext, haltRequestedConsumer, fatalErrorConsumer, simpleScratchpad);
 
         handler.issObserved(new IssNotification(1234L, IssType.CATASTROPHIC_ISS));
